+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Editor/Utils/EditorCollection.cs
@@ -19,9 +19,8 @@ using UnityEngine;
 // ReSharper disable once CheckNamespace
 
 namespace ChoyUtilities.Editor {
+    
     internal static partial class EditorCollection {
-        public const string UtilityWindow = "ChoyUtilities/";
-
         public static string ProjectFolderName => Directory.GetParent(Application.dataPath)?.Name;
 
         public static string FindPathByName(string name) {